fix: add a layout component to provide tool options
