+++ src/index.ts
@@ -1,47 +1,2 @@
-import {definePlugin} from 'sanity'
-import {ImageIcon} from '@sanity/icons'
-import type {AssetSource} from 'sanity'
-import FormBuilderTool from './components/FormBuilderTool'
-import Tool from './components/Tool'
-import mediaTag from './schemas/tag'
-import {MediaToolOptions} from '@types'
-
-const plugin = {
-  icon: ImageIcon,
-  name: 'media',
-  title: 'Media'
-}
-
-export const mediaAssetSource: AssetSource = {
-  ...plugin,
-  component: FormBuilderTool
-}
-
-export const media = definePlugin<MediaToolOptions | void>(options => ({
-  name: 'media',
-  form: {
-    file: {
-      assetSources: prev => {
-        return [...prev, mediaAssetSource]
-      }
-    },
-    image: {
-      assetSources: prev => {
-        return [...prev, mediaAssetSource]
-      }
-    }
-  },
-  schema: {
-    types: [mediaTag]
-  },
-  tools: prev => {
-    return [
-      ...prev,
-      {
-        ...plugin,
-        options,
-        component: Tool
-      }
-    ]
-  }
-}))
+export {media, mediaAssetSource} from './plugin'
+export type {MediaToolOptions} from '@types'
